Refactor add review card
